+++ src/main/java/com/everestengg/code/challenge/client/PackageDeliveryCostEstimationApp.java
@@ -1,7 +1,7 @@
 package com.everestengg.code.challenge.client;
 
 import com.everestengg.code.challenge.model.courier.PackageDeliveryCostEstimateInfo;
-import com.everestengg.code.challenge.repo.StaticOfferRepository;
+import com.everestengg.code.challenge.offer.repo.StaticOfferRepository;
 import com.everestengg.code.challenge.service.delivery.cost.estimation.PackageDeliveryCostEstimationServiceFactory;
 import com.everestengg.code.challenge.io.IoUtils;
 import com.everestengg.code.challenge.vo.InputPackage;
