+++ src/main/java/com/everestengg/code/challenge/offer/repo/CsvOfferRepository.java
@@ -1,4 +1,4 @@
-package com.everestengg.code.challenge.repo;
+package com.everestengg.code.challenge.offer.repo;
 
 import com.everestengg.code.challenge.model.csv.CsvOffer;
 import com.everestengg.code.challenge.model.csv.CsvOfferCriteria;
