+++ src/test/java/com/everestengg/code/challenge/service/DeliveryCostEstimationImplTest.java
@@ -2,7 +2,7 @@ package com.everestengg.code.challenge.service;
 
 import com.everestengg.code.challenge.exceptions.InvalidValueException;
 import com.everestengg.code.challenge.model.courier.PackageDeliveryCostEstimateInfo;
-import com.everestengg.code.challenge.repo.StaticOfferRepository;
+import com.everestengg.code.challenge.offer.repo.StaticOfferRepository;
 import com.everestengg.code.challenge.service.delivery.cost.estimation.PackageDeliveryCostEstimationImpl;
 import com.everestengg.code.challenge.vo.InputPackage;
 import com.everestengg.code.challenge.vo.Package;
